Fix #24 partial call for object members Ref #28 invocation improved slightly. Construction still pretty slow

+++ Prexonite/Commands/Core/PartialApplication/PartialCall.cs
@@ -10,7 +10,7 @@ namespace Prexonite.Commands.Core.PartialApplication
     /// </summary>
     public class PartialCall : PartialApplicationBase
     {
-        public PartialCall(sbyte[] mappings, PValue[] closedArguments) : base(mappings, closedArguments, 1)
+        public PartialCall(int[] mappings, PValue[] closedArguments) : base(mappings, closedArguments, 1)
         {
         }
